+++ sea-of-code/src/pages/home-page/components/LevelComponent.tsx
